api/webhook-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-api-python-client==2.130.0
google-auth==2.29.0
